+++ js/projections/client/config.ts
@@ -1,3 +1,4 @@
+import { DeploymentTarget } from "@fraym/proto/dist/index.freym.projections.delivery";
 import { config } from "dotenv";
 
 export interface DeliveryClientConfig {
@@ -7,6 +8,8 @@ export interface DeliveryClientConfig {
     keepaliveInterval?: number;
     // keepaliveTimeout: grpc connection keepalive ping timeout in milliseconds
     keepaliveTimeout?: number;
+    // target: optional deployment target for the client
+    deploymentTarget?: DeploymentTarget;
 }
 
 export const getEnvDeliveryConfig = (): DeliveryClientConfig => {
@@ -31,6 +34,10 @@ export const getEnvDeliveryConfig = (): DeliveryClientConfig => {
         serverAddress,
         keepaliveInterval,
         keepaliveTimeout,
+        deploymentTarget:
+            process.env.PROJECTIONS_CLIENT_DEPLOYMENT_TARGET === "green"
+                ? "DEPLOYMENT_TARGET_GREEN"
+                : "DEPLOYMENT_TARGET_BLUE",
     };
 };
 
@@ -41,9 +48,15 @@ export const useDeliveryConfigDefaults = (
         config = getEnvDeliveryConfig();
     }
 
+    const deploymentTarget =
+        process.env.PROJECTIONS_CLIENT_DEPLOYMENT_TARGET === "green"
+            ? "DEPLOYMENT_TARGET_GREEN"
+            : "DEPLOYMENT_TARGET_BLUE";
+
     return {
         serverAddress: config.serverAddress,
         keepaliveInterval: config.keepaliveInterval ?? 40 * 1000,
         keepaliveTimeout: config.keepaliveTimeout ?? 3 * 1000,
+        deploymentTarget: config.deploymentTarget ?? deploymentTarget,
     };
 };